refactoring: move ParamLoadFromJson
